backend/pyproject.toml
[tool.poetry]
name = "autobi-migrator-backend"
version = "0.1.0"
description = "Backend for AutoBI migration tool"
authors = ["Akhilesh Pothuri"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110"
uvicorn = "^0.29"
sqlalchemy = "^2.0"
psycopg2-binary = "^2.9"
python-dotenv = "^1.0"
pydantic = "^2.0" 